pyproject.toml
[project]
name = "agentic-fin-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "docling>=2.66.0",
    "edgartools>=5.8.3",
    "fastapi>=0.128.0",
    "langchain>=1.2.2",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
